requirements.txt
crewai
langchain-google-genai
python-dotenv